refactor: changed naming sim to api
